Dados do censo ensino superior INEP 2017 (IES)

Este conjunto de dados faz parte da base de dados do INEP para o censo brasileiro do ensino superior. Nesta parte temos os dados das IES Instituicoes de Ensino Superior mapeados pelo censo naquele ano. Trata-se de uma amostras das variaveis, porem o total de observacoes esta completo. Para obter dados de outros anos consulte o link do site do INEP ou a funcao rnp_get_inep_censo

References

http://inep.gov.br/microdados

Author

INEP

Examples

require(rnpextras)
str(dm_ies)
#> Classes ‘data.table’ and 'data.frame':	2448 obs. of  47 variables:
#>  $ NU_ANO_CENSO                     : int  2017 2017 2017 2017 2017 2017 2017 2017 2017 2017 ...
#>  $ CO_IES                           : int  1 2 3 4 6 10 11 13 14 18 ...
#>  $ NO_IES                           : chr  "UNIVERSIDADE FEDERAL DE MATO GROSSO" "UNIVERSIDADE DE BRASÍLIA" "UNIVERSIDADE FEDERAL DE SERGIPE" "UNIVERSIDADE FEDERAL DO AMAZONAS" ...
#>  $ SG_IES                           : chr  "UFMT" "UNB" "UFS" "UFAM" ...
#>  $ CO_MANTENEDORA                   : int  1 2 3 4 6 10 11 13 14 18 ...
#>  $ NO_MANTENEDORA                   : chr  "FUNDACAO UNIVERSIDADE FEDERAL DE MATO GROSSO" "FUNDACAO UNIVERSIDADE DE BRASILIA" "FUNDACAO UNIVERSIDADE FEDERAL DE SERGIPE" "FUNDACAO UNIVERSIDADE DO AMAZONAS" ...
#>  $ CO_REGIAO                        : int  5 5 2 1 3 4 2 4 4 4 ...
#>  $ CO_UF                            : int  51 53 28 13 31 41 26 43 43 43 ...
#>  $ CO_MUNICIPIO                     : int  5103403 5300108 2806701 1302603 3146107 4106902 2611606 4305108 4318705 4314407 ...
#>  $ QT_TEC_TOTAL                     : int  1574 3206 1429 1721 786 1275 441 986 1062 246 ...
#>  $ QT_TEC_FUNDAMENTAL_INCOMP_FEM    : int  3 0 1 12 7 1 2 8 4 3 ...
#>  $ QT_TEC_FUNDAMENTAL_INCOMP_MASC   : int  37 0 32 15 52 2 9 12 4 4 ...
#>  $ QT_TEC_FUNDAMENTAL_COMP_FEM      : int  8 15 6 1 4 4 1 22 12 2 ...
#>  $ QT_TEC_FUNDAMENTAL_COMP_MASC     : int  29 51 25 25 21 5 9 16 6 5 ...
#>  $ QT_TEC_MEDIO_FEM                 : int  110 226 120 210 36 280 59 226 314 25 ...
#>  $ QT_TEC_MEDIO_MASC                : int  189 290 167 226 101 235 73 259 283 28 ...
#>  $ QT_TEC_SUPERIOR_FEM              : int  163 299 634 207 58 309 112 215 162 76 ...
#>  $ QT_TEC_SUPERIOR_MASC             : int  150 417 432 211 104 138 91 79 58 42 ...
#>  $ QT_TEC_ESPECIALIZACAO_FEM        : int  367 770 0 386 121 150 42 90 81 30 ...
#>  $ QT_TEC_ESPECIALIZACAO_MASC       : int  270 629 0 258 118 90 32 29 29 14 ...
#>  $ QT_TEC_MESTRADO_FEM              : int  140 231 4 85 83 22 2 17 53 10 ...
#>  $ QT_TEC_MESTRADO_MASC             : int  73 198 7 57 56 21 7 8 44 5 ...
#>  $ QT_TEC_DOUTORADO_FEM             : int  22 46 1 15 15 10 0 4 7 0 ...
#>  $ QT_TEC_DOUTORADO_MASC            : int  13 34 0 13 10 8 2 1 5 2 ...
#>  $ QT_PERIODICO_ELETRONICO          : int  0 0 93536 0 0 56012 174 0 14045 38 ...
#>  $ QT_LIVRO_ELETRONICO              : int  0 1000000 248828 3212 187 24003 1210 13000 90 7232 ...
#>  $ VL_RECEITA_PROPRIA               : num  6913132 63239902 3161937 4136205 3458763 ...
#>  $ VL_RECEITA_TRANSFERENCIA         : num  9.51e+08 7.68e+07 7.48e+08 8.09e+08 4.39e+08 ...
#>  $ VL_RECEITA_OUTRA                 : num  0 48513832 5562078 16448437 4145839 ...
#>  $ VL_DESPESA_PESSOAL_DOCENTE       : num  3.70e+08 1.05e+09 2.20e+08 3.51e+08 1.67e+08 ...
#>  $ VL_DESPESA_PESSOAL_TECNICO       : num  1.46e+08 0.00 1.18e+08 0.00 8.08e+07 ...
#>  $ VL_DESPESA_PESSOAL_ENCARGO       : num  1.15e+08 1.05e+09 8.92e+07 1.08e+08 3.91e+07 ...
#>  $ VL_DESPESA_CUSTEIO               : num  1.23e+08 1.86e+08 1.23e+08 1.74e+08 3.98e+07 ...
#>  $ VL_DESPESA_INVESTIMENTO          : num  4285372 34793322 36066531 18561793 1376475 ...
#>  $ VL_DESPESA_PESQUISA              : num  5807924 5772538 3033336 1490991 748715 ...
#>  $ VL_DESPESA_OUTRA                 : num  1.94e+08 4.26e+08 1.63e+08 1.35e+08 1.17e+08 ...
#>  $ DESC_IN_ACESSO_OUTRAS_BASES      : chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_ACESSO_PORTAL_CAPES      : chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_BUSCA_INTEGRADA          : chr  "0. Não" "0. Não" "1. Sim" "0. Não" ...
#>  $ DESC_IN_CAPITAL                  : chr  "1. Sim" "1. Sim" "0. Não" "1. Sim" ...
#>  $ DESC_IN_CATALOGO_ONLINE          : chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_PARTICIPA_REDE_SOCIAL    : chr  "0. Não" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_REPOSITORIO_INSTITUCIONAL: chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_IN_SERVICO_INTERNET         : chr  "1. Sim" "1. Sim" "1. Sim" "1. Sim" ...
#>  $ DESC_TP_CATEGORIA_ADMINISTRATIVA : chr  "1. Pública Federal" "1. Pública Federal" "1. Pública Federal" "1. Pública Federal" ...
#>  $ DESC_TP_ORGANIZACAO_ACADEMICA    : chr  "1. Universidade" "1. Universidade" "1. Universidade" "1. Universidade" ...
#>  $ DESC_TP_REFERENTE                : chr  "2. Instituição" "2. Instituição" "2. Instituição" "1. Mantenedora" ...
#>  - attr(*, ".internal.selfref")=<externalptr>