Gerador de senha forte (ou absurdas)

Gerador de senha forte (ou absurdas)

Usage

gerador_de_senha(nsenhas = 1, tamanho = 16, seed = NULL, md5 = TRUE)

Arguments

nsenhas

total de senhas. Se colocar 20 saem vinte senhas

tamanho

tamanho da senha em caracteres

seed

numerico para pode repetir as senhas geradas antes. Se nao for informado usara smotragem por decimal uniforme entre 0 e 1

md5

Padrao TRUE, Devolve a senha tambem em MD5. Se FALSE devolve apenas as senhas

Value

Vetor ou data.frame com as senhas

Details

Dado o total de senhas e o tamanho delas, o programa calcula aleatoreamente as senhas com as seguintes regras: 1. Tem letras maiusculas 2. Tem letras minusculas 3. Tem numeros de 0 a 9 4. Tem caracteres especiais ex: "#","$","*",".","!","?"

Examples

# 10 senhas tamanho 8 com senha de seed de rastreamento 1234 mais MD5
gerador_de_senha(10, 8, 1234, TRUE)
#>       senha                        senha_md5
#> 1  n5]tm1fx d3db615997412e65e3410eedcd4fbfd1
#> 2  3T6]SF!5 c73e9f1eda02c0f4e9370824003c2211
#> 3  wx)]d[k0 016d6ecaf582e2a86bcd2ee26d78222c
#> 4  h0y#cS*2 ae9b3236adeba803e2bba0cddddacc58
#> 5  *6yVh#F4 ca0ea9f2035ef35eea28e27011f0b01c
#> 6  kL48n1aS 8c048b43078e6a2d7334200db47b68c1
#> 7  eYF3W$gu cb6bb193c6233b5f29ceb76005d5c725
#> 8  *Sv?m7xk 3ed3ae5f0fd2480ed533d5e43e6fe9f2
#> 9  ucm!@bZ4 9fdb2cc36842d7fd6caeaa6cd325acba
#> 10 )bxgv039 f66f1fce73a5f3e5e762f1240f3f14d8
# O mesmo sem seed
gerador_de_senha(10, 8)
#>       senha                        senha_md5
#> 1  [8?m2K@7 dd13f7464b8144e3f1d6a1d263e187a9
#> 2  6)89eY}h c93ff818786953573cca1844f6022ec0
#> 3  q3x4.ae( fad00f30de062e203265ad546f86ff53
#> 4  l2!u{RM} 5d223ee9b046b9eabb657c18e7716929
#> 5  #F6rJ]e2 3dd2cfdcb639c89a6e4a30728af97c70
#> 6  hZtf$ox8 58376b60eff083cf3f14c8e423d7a775
#> 7  rG5yRm!e ec11425fcfa47d75a426eaab51d452af
#> 8  #y0)c]?f 2f82600a68cee4c2069e5416c95a20d5
#> 9  z{i2x1?r 584386dc7b018c8d947f7db9b8d9cf85
#> 10 3!)t0X92 b975b79650907e682817a45d095ff7bf
# Agora sem seed nem md5 e com tamanho 12
gerador_de_senha(10, 12, NULL, FALSE)
#>  [1] "wPt})ZiCe83h" "[cfkASIXVPg2" "d8z!(Mg)E0R$" "q[9h5JxFsC7n" "Hb0.8JQ1AIPV"
#>  [6] "0d3$pu!Rmx]1" "2!o5d[mkcwS]" "kZKx)f6OHL#(" "{5@yF071gJA$" "hkd8*fb4T}c9"