Calculates the biggest precision and scale that occurs in a numeric vector

CREDITOS: https://stackoverflow.com/questions/42122215 The scale of a numeric is the count of decimal digits in the fractional part (to the right of the decimal point). The precision of a numeric is the total count of significant digits in the whole number, that is, the number of digits to both sides of the decimal point.

To create a suitable numeric data type in a SQL data base use the returned SQL.precision which is defined by max(precision, non.fractional.precision + scale).

Usage

numeric_precision_and_scale(x)

Arguments

x

numeric vector

Value

A list with four elements: precision (total number of significant digits in the whole number), scale (number of digits in the fractional part), non.fractional.precision (number of digits at the left side and SQL precision.

Details

NA will be counted as precision 1 and scale 0!

Examples

x <- c(0, 7654321, 54321.1234, 321.123, 321.123456789, 54321.1234, 100000000000, NA)
numeric_precision_and_scale(x)
#> $precision
#> [1] 12
#> 
#> $scale
#> [1] 9
#> 
#> $non.fractional.precision
#> [1] 12
#> 
#> $SQL.precision
#> [1] 21
#> 
numeric_precision_and_scale(c(10.0, 1.2))
#> $precision
#> [1] 2
#> 
#> $scale
#> [1] 1
#> 
#> $non.fractional.precision
#> [1] 2
#> 
#> $SQL.precision
#> [1] 3
#>