Calculo de correlacao

Esta funcao recebe como entrada uma base de dados com pelo menos duas variaveis numericas e deternina os coeficientes de correlacao de Pearson, Spearman e Kendal.

Usage

rnp_correlacao(base, digits = 4)

Arguments

base

data.frame, tibla, data.table, etc.

digits

numeros de digitos para arrendondar o valor da correlacao.

Value

Um data.frame com cinco colunas onde as duas primeiras sao as combinacoes aos pares de cada uma das variaveis e as tres ultimas sao as correlacoes de Pearson, Spearman e Kendal respectivamente.

Details

E possivel que existam variaveis nao numerias em sua base de dados. Neste caso, elas serao eliminadas. Para mais detalhes veja cor.

Author

LOPES, J. E.

Examples

rnp_correlacao(mtcars)
#>       x    y pearson spearman kendall
#> 43  mpg   wt -0.8677  -0.8864 -0.7278
#> 36  mpg  cyl -0.8522  -0.9108 -0.7953
#> 37  mpg disp -0.8476  -0.9089 -0.7681
#> 40  mpg   hp -0.7762  -0.8947 -0.7428
#> 38  mpg drat  0.6812   0.6515  0.4645
#> 42  mpg   vs  0.6640   0.7066  0.5897
#> 34  mpg   am  0.5998   0.5620  0.4690
#> 35  mpg carb -0.5509  -0.6575 -0.5044
#> 39  mpg gear  0.4803   0.5428  0.4332
#> 41  mpg qsec  0.4187   0.4669  0.3154
#> 5   cyl disp  0.9020   0.9277  0.8144
#> 8   cyl   hp  0.8324   0.9018  0.7852
#> 10  cyl   vs -0.8108  -0.8138 -0.7710
#> 11  cyl   wt  0.7825   0.8577  0.7283
#> 6   cyl drat -0.6999  -0.6789 -0.5513
#> 9   cyl qsec -0.5912  -0.5724 -0.4490
#> 4   cyl carb  0.5270   0.5801  0.4654
#> 3   cyl   am -0.5226  -0.5221 -0.4946
#> 7   cyl gear -0.4927  -0.5643 -0.5125
#> 19 disp   wt  0.8880   0.8977  0.7434
#> 16 disp   hp  0.7909   0.8510  0.6660
#> 18 disp   vs -0.7104  -0.7237 -0.6033
#> 14 disp drat -0.7102  -0.6836 -0.4990
#> 12 disp   am -0.5912  -0.6241 -0.5203
#> 15 disp gear -0.5556  -0.5945 -0.4760
#> 17 disp qsec -0.4337  -0.4598 -0.3008
#> 13 disp carb  0.3950   0.5398  0.4137
#> 28   hp carb  0.7498   0.7334  0.5960
#> 32   hp   vs -0.7231  -0.7516 -0.6306
#> 31   hp qsec -0.7082  -0.6666 -0.4729
#> 33   hp   wt  0.6587   0.7747  0.6113
#> 29   hp drat -0.4488  -0.5201 -0.3826
#> 27   hp   am -0.2432  -0.3623 -0.3040
#> 30   hp gear -0.1257  -0.3314 -0.2794
#> 20 drat   am  0.7127   0.6866  0.5755
#> 25 drat   wt -0.7124  -0.7504 -0.5471
#> 22 drat gear  0.6996   0.7448  0.5839
#> 24 drat   vs  0.4403   0.4475  0.3751
#> 23 drat qsec  0.0912   0.0919  0.0327
#> 21 drat carb -0.0908  -0.1252 -0.0954
#> 51   wt   am -0.6925  -0.7377 -0.6139
#> 53   wt gear -0.5833  -0.6761 -0.5436
#> 55   wt   vs -0.5549  -0.5870 -0.4885
#> 52   wt carb  0.4276   0.4998  0.3714
#> 54   wt qsec -0.1747  -0.2254 -0.1420
#> 47 qsec   vs  0.7445   0.7916  0.6575
#> 45 qsec carb -0.6562  -0.6587 -0.5064
#> 44 qsec   am -0.2299  -0.2033 -0.1689
#> 46 qsec gear -0.2127  -0.1482 -0.0913
#> 49   vs carb -0.5696  -0.6337 -0.5769
#> 50   vs gear  0.2060   0.2827  0.2697
#> 48   vs   am  0.1683   0.1683  0.1683
#> 2    am gear  0.7941   0.8077  0.7708
#> 1    am carb  0.0575  -0.0644 -0.0586
#> 26 gear carb  0.2741   0.1149  0.0980