Detecta tipos e formata para o SQL

Esta funcao e um trator que vasculha todos os dados do objeto informado buscando identificar os tipos de dados e seus tamanhos maximos com base no total de caracteres retoando a formatacao ideal para o SQL.

Usage

rnp_db_types(
  con,
  data,
  sample_size = Inf,
  parallel = FALSE,
  aloc_cpu = 0.2,
  table = FALSE,
  verbose = TRUE,
  max_mem_size = 2 * 1024^3
)

Arguments

con

Conexao criada via dbConnect.

data

Objeto para obter tipos dos dados.

sample_size

Para detectar tipos com base em uma amostra definida pelo usuario. Aceita quantidade de linhas (valor entre 1 e nrow(data)) ou percentual da base (valor entre 0.01 e 1.00).

parallel

Se TRUE, ativa processamento em paralelo via future. Veja mais em plan.

aloc_cpu

Percentual de toda a CPU utilizada no processamento em paralelo. Valor entre 0.01 e 0.80. Ex. se tiver 10 cores e aloc_cpu = 0.10, roda apenas com um core.

table

Se TRUE, retorna tabela com dados dos tipos detecados para o objeto data.

verbose

Se TRUE, exibe historico de execussao.

max_mem_size

Total maximo de memoria fisica alocada para trabalhar em paralelo. Padrao 2Gb (2*1024^3). Veja future.options para mais detalhes.

Details

O objetivo principal dela e informar para a funcao rnp_db_write ou qualquer outra da familia DBI que possua o argumento field.types ou data.types, por exemplo dbWriteTable, o formato correto para o comando de CREATE TABLE do SQL. Isso e uma mao na roda para que o SQL crie a tabela ja tipada com os formatos corretos evitando uso excessivo de banco e posterior trabalho de tipagem em SQL que nem sempre e amigavel. Se os dados sao numericos o R busca a quantidade de caracteres antes da virgula (escala) e o total depois da virgula (precisao) e cria a formatacao correspondente ao SQL Server. Se os dados sao do tipo charactere com mais de 7999 caracteres o formato SQL sera VARCHAR(MAX). O processo de scan dos dados pode ser lento em tabelas grandes em linhas ou colunas. Para tentar melhorar a performance nestes casos, foi inserido na funcao a possibilidade de rodar em paralelo com apoio do pacote future. Veja multiprocess para mais detalhes.

Examples

if (FALSE) {
rnp_db_types(con, mtcars)
rnp_db_types(con, mtcars, parallel = TRUE, table = TRUE)
}