Tabela de frequencias simples

Calcula estatisticas descritivas de variaveis categoricas como caractere e fatores. Tambem funciona com numericas com valores repetidos. Como resultado tem-se frequencias e percentuais simples e acumulados de cada categoria. Se receber um data.frame tenta calcular as estatisticas para todas as variaveis da mesma. Nao existe deduplicacao dos dados.

Usage

rnp_freq(x, digits = 4, percent = FALSE, sortclass = TRUE)

Arguments

x

varivetor de entrada

digits

arredondar as frequencias para...

percent

se TRUE retorna os percentuais formatados como %

sortclass

ordenacao das classes.

Value

data.frame com as estaisticas frequencia absoluta (fa), relativa (fr) absoluta acumulada (faa) e relativa acumulada (fra)

Author

LOPES, J. E.

Examples

rnp_freq(mtcars$cyl)
#>   classe fa     fr faa    fra
#> 1      4 11 0.3438  11 0.3438
#> 2      6  7 0.2188  18 0.5625
#> 3      8 14 0.4375  32 1.0000
rnp_freq(mtcars$cyl, percent = TRUE)
#>   classe fa      fr faa     fra
#> 1      4 11 34.375%  11 34.375%
#> 2      6  7 21.875%  18  56.25%
#> 3      8 14  43.75%  32    100%
rnp_freq(mtcars, percent = TRUE)
#>     variavel  classe fa      fr faa     fra
#> 1        mpg  10.400  2   6.25%   2   6.25%
#> 2        mpg  13.300  1  3.125%   3  9.375%
#> 3        mpg  14.300  1  3.125%   4   12.5%
#> 4        mpg  14.700  1  3.125%   5 15.625%
#> 5        mpg  15.000  1  3.125%   6  18.75%
#> 6        mpg  15.200  2   6.25%   8     25%
#> 7        mpg  15.500  1  3.125%   9 28.125%
#> 8        mpg  15.800  1  3.125%  10  31.25%
#> 9        mpg  16.400  1  3.125%  11 34.375%
#> 10       mpg  17.300  1  3.125%  12   37.5%
#> 11       mpg  17.800  1  3.125%  13 40.625%
#> 12       mpg  18.100  1  3.125%  14  43.75%
#> 13       mpg  18.700  1  3.125%  15 46.875%
#> 14       mpg  19.200  2   6.25%  17 53.125%
#> 15       mpg  19.700  1  3.125%  18  56.25%
#> 16       mpg  21.000  2   6.25%  20   62.5%
#> 17       mpg  21.400  2   6.25%  22  68.75%
#> 18       mpg  21.500  1  3.125%  23 71.875%
#> 19       mpg  22.800  2   6.25%  25 78.125%
#> 20       mpg  24.400  1  3.125%  26  81.25%
#> 21       mpg  26.000  1  3.125%  27 84.375%
#> 22       mpg  27.300  1  3.125%  28   87.5%
#> 23       mpg  30.400  2   6.25%  30  93.75%
#> 24       mpg  32.400  1  3.125%  31 96.875%
#> 25       mpg  33.900  1  3.125%  32    100%
#> 26       cyl   4.000 11 34.375%  11 34.375%
#> 27       cyl   6.000  7 21.875%  18  56.25%
#> 28       cyl   8.000 14  43.75%  32    100%
#> 29      disp  71.100  1  3.125%   1  3.125%
#> 30      disp  75.700  1  3.125%   2   6.25%
#> 31      disp  78.700  1  3.125%   3  9.375%
#> 32      disp  79.000  1  3.125%   4   12.5%
#> 33      disp  95.100  1  3.125%   5 15.625%
#> 34      disp 108.000  1  3.125%   6  18.75%
#> 35      disp 120.100  1  3.125%   7 21.875%
#> 36      disp 120.300  1  3.125%   8     25%
#> 37      disp 121.000  1  3.125%   9 28.125%
#> 38      disp 140.800  1  3.125%  10  31.25%
#> 39      disp 145.000  1  3.125%  11 34.375%
#> 40      disp 146.700  1  3.125%  12   37.5%
#> 41      disp 160.000  2   6.25%  14  43.75%
#> 42      disp 167.600  2   6.25%  16     50%
#> 43      disp 225.000  1  3.125%  17 53.125%
#> 44      disp 258.000  1  3.125%  18  56.25%
#> 45      disp 275.800  3  9.375%  21 65.625%
#> 46      disp 301.000  1  3.125%  22  68.75%
#> 47      disp 304.000  1  3.125%  23 71.875%
#> 48      disp 318.000  1  3.125%  24     75%
#> 49      disp 350.000  1  3.125%  25 78.125%
#> 50      disp 351.000  1  3.125%  26  81.25%
#> 51      disp 360.000  2   6.25%  28   87.5%
#> 52      disp 400.000  1  3.125%  29 90.625%
#> 53      disp 440.000  1  3.125%  30  93.75%
#> 54      disp 460.000  1  3.125%  31 96.875%
#> 55      disp 472.000  1  3.125%  32    100%
#> 56        hp  52.000  1  3.125%   1  3.125%
#> 57        hp  62.000  1  3.125%   2   6.25%
#> 58        hp  65.000  1  3.125%   3  9.375%
#> 59        hp  66.000  2   6.25%   5 15.625%
#> 60        hp  91.000  1  3.125%   6  18.75%
#> 61        hp  93.000  1  3.125%   7 21.875%
#> 62        hp  95.000  1  3.125%   8     25%
#> 63        hp  97.000  1  3.125%   9 28.125%
#> 64        hp 105.000  1  3.125%  10  31.25%
#> 65        hp 109.000  1  3.125%  11 34.375%
#> 66        hp 110.000  3  9.375%  14  43.75%
#> 67        hp 113.000  1  3.125%  15 46.875%
#> 68        hp 123.000  2   6.25%  17 53.125%
#> 69        hp 150.000  2   6.25%  19 59.375%
#> 70        hp 175.000  3  9.375%  22  68.75%
#> 71        hp 180.000  3  9.375%  25 78.125%
#> 72        hp 205.000  1  3.125%  26  81.25%
#> 73        hp 215.000  1  3.125%  27 84.375%
#> 74        hp 230.000  1  3.125%  28   87.5%
#> 75        hp 245.000  2   6.25%  30  93.75%
#> 76        hp 264.000  1  3.125%  31 96.875%
#> 77        hp 335.000  1  3.125%  32    100%
#> 78      drat   2.760  2   6.25%   2   6.25%
#> 79      drat   2.930  1  3.125%   3  9.375%
#> 80      drat   3.000  1  3.125%   4   12.5%
#> 81      drat   3.070  3  9.375%   7 21.875%
#> 82      drat   3.080  2   6.25%   9 28.125%
#> 83      drat   3.150  2   6.25%  11 34.375%
#> 84      drat   3.210  1  3.125%  12   37.5%
#> 85      drat   3.230  1  3.125%  13 40.625%
#> 86      drat   3.540  1  3.125%  14  43.75%
#> 87      drat   3.620  1  3.125%  15 46.875%
#> 88      drat   3.690  1  3.125%  16     50%
#> 89      drat   3.700  1  3.125%  17 53.125%
#> 90      drat   3.730  1  3.125%  18  56.25%
#> 91      drat   3.770  1  3.125%  19 59.375%
#> 92      drat   3.850  1  3.125%  20   62.5%
#> 93      drat   3.900  2   6.25%  22  68.75%
#> 94      drat   3.920  3  9.375%  25 78.125%
#> 95      drat   4.080  2   6.25%  27 84.375%
#> 96      drat   4.110  1  3.125%  28   87.5%
#> 97      drat   4.220  2   6.25%  30  93.75%
#> 98      drat   4.430  1  3.125%  31 96.875%
#> 99      drat   4.930  1  3.125%  32    100%
#> 100       wt   1.513  1  3.125%   1  3.125%
#> 101       wt   1.615  1  3.125%   2   6.25%
#> 102       wt   1.835  1  3.125%   3  9.375%
#> 103       wt   1.935  1  3.125%   4   12.5%
#> 104       wt   2.140  1  3.125%   5 15.625%
#> 105       wt   2.200  1  3.125%   6  18.75%
#> 106       wt   2.320  1  3.125%   7 21.875%
#> 107       wt   2.465  1  3.125%   8     25%
#> 108       wt   2.620  1  3.125%   9 28.125%
#> 109       wt   2.770  1  3.125%  10  31.25%
#> 110       wt   2.780  1  3.125%  11 34.375%
#> 111       wt   2.875  1  3.125%  12   37.5%
#> 112       wt   3.150  1  3.125%  13 40.625%
#> 113       wt   3.170  1  3.125%  14  43.75%
#> 114       wt   3.190  1  3.125%  15 46.875%
#> 115       wt   3.215  1  3.125%  16     50%
#> 116       wt   3.435  1  3.125%  17 53.125%
#> 117       wt   3.440  3  9.375%  20   62.5%
#> 118       wt   3.460  1  3.125%  21 65.625%
#> 119       wt   3.520  1  3.125%  22  68.75%
#> 120       wt   3.570  2   6.25%  24     75%
#> 121       wt   3.730  1  3.125%  25 78.125%
#> 122       wt   3.780  1  3.125%  26  81.25%
#> 123       wt   3.840  1  3.125%  27 84.375%
#> 124       wt   3.845  1  3.125%  28   87.5%
#> 125       wt   4.070  1  3.125%  29 90.625%
#> 126       wt   5.250  1  3.125%  30  93.75%
#> 127       wt   5.345  1  3.125%  31 96.875%
#> 128       wt   5.424  1  3.125%  32    100%
#> 129     qsec  14.500  1  3.125%   1  3.125%
#> 130     qsec  14.600  1  3.125%   2   6.25%
#> 131     qsec  15.410  1  3.125%   3  9.375%
#> 132     qsec  15.500  1  3.125%   4   12.5%
#> 133     qsec  15.840  1  3.125%   5 15.625%
#> 134     qsec  16.460  1  3.125%   6  18.75%
#> 135     qsec  16.700  1  3.125%   7 21.875%
#> 136     qsec  16.870  1  3.125%   8     25%
#> 137     qsec  16.900  1  3.125%   9 28.125%
#> 138     qsec  17.020  2   6.25%  11 34.375%
#> 139     qsec  17.050  1  3.125%  12   37.5%
#> 140     qsec  17.300  1  3.125%  13 40.625%
#> 141     qsec  17.400  1  3.125%  14  43.75%
#> 142     qsec  17.420  1  3.125%  15 46.875%
#> 143     qsec  17.600  1  3.125%  16     50%
#> 144     qsec  17.820  1  3.125%  17 53.125%
#> 145     qsec  17.980  1  3.125%  18  56.25%
#> 146     qsec  18.000  1  3.125%  19 59.375%
#> 147     qsec  18.300  1  3.125%  20   62.5%
#> 148     qsec  18.520  1  3.125%  21 65.625%
#> 149     qsec  18.600  1  3.125%  22  68.75%
#> 150     qsec  18.610  1  3.125%  23 71.875%
#> 151     qsec  18.900  2   6.25%  25 78.125%
#> 152     qsec  19.440  1  3.125%  26  81.25%
#> 153     qsec  19.470  1  3.125%  27 84.375%
#> 154     qsec  19.900  1  3.125%  28   87.5%
#> 155     qsec  20.000  1  3.125%  29 90.625%
#> 156     qsec  20.010  1  3.125%  30  93.75%
#> 157     qsec  20.220  1  3.125%  31 96.875%
#> 158     qsec  22.900  1  3.125%  32    100%
#> 159       vs   0.000 18  56.25%  18  56.25%
#> 160       vs   1.000 14  43.75%  32    100%
#> 161       am   0.000 19 59.375%  19 59.375%
#> 162       am   1.000 13 40.625%  32    100%
#> 163     gear   3.000 15 46.875%  15 46.875%
#> 164     gear   4.000 12   37.5%  27 84.375%
#> 165     gear   5.000  5 15.625%  32    100%
#> 166     carb   1.000  7 21.875%   7 21.875%
#> 167     carb   2.000 10  31.25%  17 53.125%
#> 168     carb   3.000  3  9.375%  20   62.5%
#> 169     carb   4.000 10  31.25%  30  93.75%
#> 170     carb   6.000  1  3.125%  31 96.875%
#> 171     carb   8.000  1  3.125%  32    100%