Tabela de frequencias simples

Calcula estatisticas descritivas de variaveis categoricas como caractere e fatores. Tambem funciona com numericas com valores repetidos. Como resultado tem-se frequencias e percentuais simples e acumulados de cada categoria. Se receber um data.frame ou lista tenta calcular as estatisticas para todas as variaveis da mesma. Nao existe deduplicacao dos dados.

Usage

rnp_freq2(x, sortclass = TRUE)

Arguments

x

variavel de entrada

sortclass

ordenacao das classes.

Value

data.frame com as estaisticas frequencia absoluta (fa), relativa (fr) absoluta acumulada (faa) e relativa acumulada (fra)

Author

LOPES, J. E.

Examples

rnp_freq(mtcars$cyl)
#>   classe fa     fr faa    fra
#> 1      4 11 0.3438  11 0.3438
#> 2      6  7 0.2188  18 0.5625
#> 3      8 14 0.4375  32 1.0000
rnp_freq(mtcars)
#>     variavel  classe fa     fr faa    fra
#> 1        mpg  10.400  2 0.0625   2 0.0625
#> 2        mpg  13.300  1 0.0312   3 0.0938
#> 3        mpg  14.300  1 0.0312   4 0.1250
#> 4        mpg  14.700  1 0.0312   5 0.1562
#> 5        mpg  15.000  1 0.0312   6 0.1875
#> 6        mpg  15.200  2 0.0625   8 0.2500
#> 7        mpg  15.500  1 0.0312   9 0.2812
#> 8        mpg  15.800  1 0.0312  10 0.3125
#> 9        mpg  16.400  1 0.0312  11 0.3438
#> 10       mpg  17.300  1 0.0312  12 0.3750
#> 11       mpg  17.800  1 0.0312  13 0.4062
#> 12       mpg  18.100  1 0.0312  14 0.4375
#> 13       mpg  18.700  1 0.0312  15 0.4688
#> 14       mpg  19.200  2 0.0625  17 0.5312
#> 15       mpg  19.700  1 0.0312  18 0.5625
#> 16       mpg  21.000  2 0.0625  20 0.6250
#> 17       mpg  21.400  2 0.0625  22 0.6875
#> 18       mpg  21.500  1 0.0312  23 0.7188
#> 19       mpg  22.800  2 0.0625  25 0.7812
#> 20       mpg  24.400  1 0.0312  26 0.8125
#> 21       mpg  26.000  1 0.0312  27 0.8438
#> 22       mpg  27.300  1 0.0312  28 0.8750
#> 23       mpg  30.400  2 0.0625  30 0.9375
#> 24       mpg  32.400  1 0.0312  31 0.9688
#> 25       mpg  33.900  1 0.0312  32 1.0000
#> 26       cyl   4.000 11 0.3438  11 0.3438
#> 27       cyl   6.000  7 0.2188  18 0.5625
#> 28       cyl   8.000 14 0.4375  32 1.0000
#> 29      disp  71.100  1 0.0312   1 0.0312
#> 30      disp  75.700  1 0.0312   2 0.0625
#> 31      disp  78.700  1 0.0312   3 0.0938
#> 32      disp  79.000  1 0.0312   4 0.1250
#> 33      disp  95.100  1 0.0312   5 0.1562
#> 34      disp 108.000  1 0.0312   6 0.1875
#> 35      disp 120.100  1 0.0312   7 0.2188
#> 36      disp 120.300  1 0.0312   8 0.2500
#> 37      disp 121.000  1 0.0312   9 0.2812
#> 38      disp 140.800  1 0.0312  10 0.3125
#> 39      disp 145.000  1 0.0312  11 0.3438
#> 40      disp 146.700  1 0.0312  12 0.3750
#> 41      disp 160.000  2 0.0625  14 0.4375
#> 42      disp 167.600  2 0.0625  16 0.5000
#> 43      disp 225.000  1 0.0312  17 0.5312
#> 44      disp 258.000  1 0.0312  18 0.5625
#> 45      disp 275.800  3 0.0938  21 0.6562
#> 46      disp 301.000  1 0.0312  22 0.6875
#> 47      disp 304.000  1 0.0312  23 0.7188
#> 48      disp 318.000  1 0.0312  24 0.7500
#> 49      disp 350.000  1 0.0312  25 0.7812
#> 50      disp 351.000  1 0.0312  26 0.8125
#> 51      disp 360.000  2 0.0625  28 0.8750
#> 52      disp 400.000  1 0.0312  29 0.9062
#> 53      disp 440.000  1 0.0312  30 0.9375
#> 54      disp 460.000  1 0.0312  31 0.9688
#> 55      disp 472.000  1 0.0312  32 1.0000
#> 56        hp  52.000  1 0.0312   1 0.0312
#> 57        hp  62.000  1 0.0312   2 0.0625
#> 58        hp  65.000  1 0.0312   3 0.0938
#> 59        hp  66.000  2 0.0625   5 0.1562
#> 60        hp  91.000  1 0.0312   6 0.1875
#> 61        hp  93.000  1 0.0312   7 0.2188
#> 62        hp  95.000  1 0.0312   8 0.2500
#> 63        hp  97.000  1 0.0312   9 0.2812
#> 64        hp 105.000  1 0.0312  10 0.3125
#> 65        hp 109.000  1 0.0312  11 0.3438
#> 66        hp 110.000  3 0.0938  14 0.4375
#> 67        hp 113.000  1 0.0312  15 0.4688
#> 68        hp 123.000  2 0.0625  17 0.5312
#> 69        hp 150.000  2 0.0625  19 0.5938
#> 70        hp 175.000  3 0.0938  22 0.6875
#> 71        hp 180.000  3 0.0938  25 0.7812
#> 72        hp 205.000  1 0.0312  26 0.8125
#> 73        hp 215.000  1 0.0312  27 0.8438
#> 74        hp 230.000  1 0.0312  28 0.8750
#> 75        hp 245.000  2 0.0625  30 0.9375
#> 76        hp 264.000  1 0.0312  31 0.9688
#> 77        hp 335.000  1 0.0312  32 1.0000
#> 78      drat   2.760  2 0.0625   2 0.0625
#> 79      drat   2.930  1 0.0312   3 0.0938
#> 80      drat   3.000  1 0.0312   4 0.1250
#> 81      drat   3.070  3 0.0938   7 0.2188
#> 82      drat   3.080  2 0.0625   9 0.2812
#> 83      drat   3.150  2 0.0625  11 0.3438
#> 84      drat   3.210  1 0.0312  12 0.3750
#> 85      drat   3.230  1 0.0312  13 0.4062
#> 86      drat   3.540  1 0.0312  14 0.4375
#> 87      drat   3.620  1 0.0312  15 0.4688
#> 88      drat   3.690  1 0.0312  16 0.5000
#> 89      drat   3.700  1 0.0312  17 0.5312
#> 90      drat   3.730  1 0.0312  18 0.5625
#> 91      drat   3.770  1 0.0312  19 0.5938
#> 92      drat   3.850  1 0.0312  20 0.6250
#> 93      drat   3.900  2 0.0625  22 0.6875
#> 94      drat   3.920  3 0.0938  25 0.7812
#> 95      drat   4.080  2 0.0625  27 0.8438
#> 96      drat   4.110  1 0.0312  28 0.8750
#> 97      drat   4.220  2 0.0625  30 0.9375
#> 98      drat   4.430  1 0.0312  31 0.9688
#> 99      drat   4.930  1 0.0312  32 1.0000
#> 100       wt   1.513  1 0.0312   1 0.0312
#> 101       wt   1.615  1 0.0312   2 0.0625
#> 102       wt   1.835  1 0.0312   3 0.0938
#> 103       wt   1.935  1 0.0312   4 0.1250
#> 104       wt   2.140  1 0.0312   5 0.1562
#> 105       wt   2.200  1 0.0312   6 0.1875
#> 106       wt   2.320  1 0.0312   7 0.2188
#> 107       wt   2.465  1 0.0312   8 0.2500
#> 108       wt   2.620  1 0.0312   9 0.2812
#> 109       wt   2.770  1 0.0312  10 0.3125
#> 110       wt   2.780  1 0.0312  11 0.3438
#> 111       wt   2.875  1 0.0312  12 0.3750
#> 112       wt   3.150  1 0.0312  13 0.4062
#> 113       wt   3.170  1 0.0312  14 0.4375
#> 114       wt   3.190  1 0.0312  15 0.4688
#> 115       wt   3.215  1 0.0312  16 0.5000
#> 116       wt   3.435  1 0.0312  17 0.5312
#> 117       wt   3.440  3 0.0938  20 0.6250
#> 118       wt   3.460  1 0.0312  21 0.6562
#> 119       wt   3.520  1 0.0312  22 0.6875
#> 120       wt   3.570  2 0.0625  24 0.7500
#> 121       wt   3.730  1 0.0312  25 0.7812
#> 122       wt   3.780  1 0.0312  26 0.8125
#> 123       wt   3.840  1 0.0312  27 0.8438
#> 124       wt   3.845  1 0.0312  28 0.8750
#> 125       wt   4.070  1 0.0312  29 0.9062
#> 126       wt   5.250  1 0.0312  30 0.9375
#> 127       wt   5.345  1 0.0312  31 0.9688
#> 128       wt   5.424  1 0.0312  32 1.0000
#> 129     qsec  14.500  1 0.0312   1 0.0312
#> 130     qsec  14.600  1 0.0312   2 0.0625
#> 131     qsec  15.410  1 0.0312   3 0.0938
#> 132     qsec  15.500  1 0.0312   4 0.1250
#> 133     qsec  15.840  1 0.0312   5 0.1562
#> 134     qsec  16.460  1 0.0312   6 0.1875
#> 135     qsec  16.700  1 0.0312   7 0.2188
#> 136     qsec  16.870  1 0.0312   8 0.2500
#> 137     qsec  16.900  1 0.0312   9 0.2812
#> 138     qsec  17.020  2 0.0625  11 0.3438
#> 139     qsec  17.050  1 0.0312  12 0.3750
#> 140     qsec  17.300  1 0.0312  13 0.4062
#> 141     qsec  17.400  1 0.0312  14 0.4375
#> 142     qsec  17.420  1 0.0312  15 0.4688
#> 143     qsec  17.600  1 0.0312  16 0.5000
#> 144     qsec  17.820  1 0.0312  17 0.5312
#> 145     qsec  17.980  1 0.0312  18 0.5625
#> 146     qsec  18.000  1 0.0312  19 0.5938
#> 147     qsec  18.300  1 0.0312  20 0.6250
#> 148     qsec  18.520  1 0.0312  21 0.6562
#> 149     qsec  18.600  1 0.0312  22 0.6875
#> 150     qsec  18.610  1 0.0312  23 0.7188
#> 151     qsec  18.900  2 0.0625  25 0.7812
#> 152     qsec  19.440  1 0.0312  26 0.8125
#> 153     qsec  19.470  1 0.0312  27 0.8438
#> 154     qsec  19.900  1 0.0312  28 0.8750
#> 155     qsec  20.000  1 0.0312  29 0.9062
#> 156     qsec  20.010  1 0.0312  30 0.9375
#> 157     qsec  20.220  1 0.0312  31 0.9688
#> 158     qsec  22.900  1 0.0312  32 1.0000
#> 159       vs   0.000 18 0.5625  18 0.5625
#> 160       vs   1.000 14 0.4375  32 1.0000
#> 161       am   0.000 19 0.5938  19 0.5938
#> 162       am   1.000 13 0.4062  32 1.0000
#> 163     gear   3.000 15 0.4688  15 0.4688
#> 164     gear   4.000 12 0.3750  27 0.8438
#> 165     gear   5.000  5 0.1562  32 1.0000
#> 166     carb   1.000  7 0.2188   7 0.2188
#> 167     carb   2.000 10 0.3125  17 0.5312
#> 168     carb   3.000  3 0.0938  20 0.6250
#> 169     carb   4.000 10 0.3125  30 0.9375
#> 170     carb   6.000  1 0.0312  31 0.9688
#> 171     carb   8.000  1 0.0312  32 1.0000