Obtem as classes das variaveis censo INEP

Obtem as classes das variaveis do diconario de dados do censo INEP.

Usage

rnp_get_classes_inep(caminho, aba = 1, pula_linha = 1, retorna_lista = FALSE)

Arguments

caminho

caminho do arquivo Excel do dicionario de dados do INEP

aba

aba da planilha correspondente aos dados que deseja

pula_linha

quantidade de linhas que deseja pular

retorna_lista

TRUE se quer obter uma lista de data.frames, sendo um para cada variavel ou base ja agregada.

Details

Exige que passe o caminho para o dicionario de dados em excel e o nome da aba correspondente ou o numero da mesma, ex. 1 = primeira aba e assim por diante. A funcao faz a leitura dos dados utilizando 'read_excel' e retorna um data.frame com 4 aolunas contendo ordem, nome, categoria e descricao para todas as variaveis da base informada. OBS.: Testado apenas nos dados do censo de 2017.

Author

LOPES, J. E.

Examples

if (FALSE) {
nn <- c("DM_CURSO","DM_IES","DM_LOCAL_OFERTA","DM_DOCENTE")
L <- plyr::llply(nn[1], function(base){
  classes <- rnp_get_classes_inep(caminho = "Dicionario_de_Dados.xlsx",
                                  aba = base, retorna_lista = FALSE)
  base <- rnp_read(base = paste0("Dados/CSV/", base, ".CSV"),
                   sep = "|",
                   dec = ".",
                   header = TRUE,
                   encoding = "Latin-1",
                   verbose = FALSE,
                   showProgress = FALSE)
  oo <- rnp_aplica_classes(base = base, classes = classes)
  return(oo)
}, .progress = "text")
names(L) <- nn
}