Calcula medias de varios tipos

Esta funcao e uma chamada master paras as funcoes media_aritmetica, media_geometrica e media_harmonica. Serve para calcular médias de vetores numéricos e de matrizes/data.frames formados por colunas numericas.

Usage

rnp_media(x, peso = NULL, remove.na = TRUE)

Arguments

x

Vetor, matrix ou data.frame de dados numericos

peso

Opcional. Vetor, matrix ou data.frame de dados numericos com pesos com as mesmas caracteristcias de comprimento ou dimensoes do argumento x

remove.na

Se TRUE remove os NA's, caso existam nos dados

Value

Vetor ou data.frame com as medias de uma ou mais variaveis. No caso de data.frames, a saida e um data.frame com quatro colunas, sendo uma o nome das variaveis e as outras tres as colunas de medias aritmetica, geometrica e harmonica respectivamente. No caso de vetor numerico, a saida e um data.frame simples com tres colunas e uma linha contendo as medias.

Details

Atraves desta funcao e possivel calcular medias aritmeticas, geometricas e harmonicas simples e ponderadas. Se for passado uma matriz ela varre tdoas as colunas e determina as medias para cada variavel. Se for passado argumento peso, as medias sao ponderadas pelos seus respectivos pesos. No caso de matrizes, a matriz de peso precisa ter as mesmas dimensoes da matriz de origem e o mesmo tipo de dados. Quando existir peso, o nome das colunas do data.frame serao nomeadas como: media_ap, media_gp e media_hp. Caso contrario, media_a, media_g e media_h.

Author

LOPES, J. E.

Examples

require(rnpextras)
# Exemplos com vetores
x <- mtcars$mpg;x[3:5] <- NA
p <- mtcars$cyl;p[5:7] <- NA
rnp_media(x = x, peso = NULL, remove.na = TRUE)
#>   media_a  media_g  media_h
#> 1      20 19.08737 18.22516
rnp_media(x = x, peso = NULL, remove.na = FALSE)
#>   media_a media_g media_h
#> 1      NA      NA      NA
rnp_media(x = x, peso = p, remove.na = TRUE)
#>   media_ap media_gp media_hp
#> 1 18.68313 17.86203 17.10918
rnp_media(x = x, peso = p, remove.na = FALSE)
#>   media_ap media_gp media_hp
#> 1       NA       NA       NA
# Exemplos com matrizes (de mesmo tipo)
X <- mtcars
Y <- data.frame(matrix(runif(prod(dim(mtcars)), 5, 10), nrow = nrow(mtcars), ncol = ncol(mtcars)));
class(X); class(Y)
#> [1] "data.frame"
#> [1] "data.frame"
names(Y) <- names(X)
dim(X) == dim(Y)
#> [1] TRUE TRUE
rnp_media(X)
#>    variavel    media_a    media_g    media_h
#> 1       mpg  20.090625  19.250064  18.440917
#> 2       cyl   6.187500   5.919439   5.647059
#> 3      disp 230.721875 197.321758 166.799389
#> 4        hp 146.687500 131.883679 118.228892
#> 5      drat   3.596563   3.558394   3.520704
#> 6        wt   3.217250   3.070189   2.918263
#> 7      qsec  17.848750  17.763115  17.678200
#> 8        vs   0.437500   0.000000   0.000000
#> 9        am   0.406250   0.000000   0.000000
#> 10     gear   3.687500   3.619405   3.555556
#> 11     carb   2.812500   2.395987   2.026385
rnp_media(X, Y)
#>    variavel    media_ap   media_gp   media_hp
#> 1       mpg  20.0145592  19.161087  18.345717
#> 2       cyl   6.1505342   5.882528   5.612137
#> 3      disp 228.1158450 193.911486 163.509958
#> 4        hp 152.4135772 137.334507 123.067665
#> 5      drat   3.5811975   3.541643   3.502692
#> 6        wt   3.2357518   3.085000   2.930360
#> 7      qsec  17.8686927  17.783891  17.700130
#> 8        vs   0.4259112   0.000000   0.000000
#> 9        am   0.3799267   0.000000   0.000000
#> 10     gear   3.7163845   3.647248   3.581960
#> 11     carb   2.8472200   2.407007   2.019011
# Com dados reais da base rnpextras::dm_ies
X <- as.data.frame(rnpextras::dm_ies)
X <- X[,c(sapply(X, class) %in% c("integer", "numeric"))]
rnp_media(X[,-c(1:6)])
#>                          variavel      media_a      media_g   media_h
#> 1                    QT_TEC_TOTAL 1.680535e+02 3.808844e+01  15.10744
#> 2   QT_TEC_FUNDAMENTAL_INCOMP_FEM 2.196078e+00 0.000000e+00   0.00000
#> 3  QT_TEC_FUNDAMENTAL_INCOMP_MASC 3.773284e+00 0.000000e+00   0.00000
#> 4     QT_TEC_FUNDAMENTAL_COMP_FEM 3.393382e+00 0.000000e+00   0.00000
#> 5    QT_TEC_FUNDAMENTAL_COMP_MASC 5.010212e+00 0.000000e+00   0.00000
#> 6                QT_TEC_MEDIO_FEM 2.725613e+01 0.000000e+00   0.00000
#> 7               QT_TEC_MEDIO_MASC 2.750654e+01 0.000000e+00   0.00000
#> 8             QT_TEC_SUPERIOR_FEM 2.786928e+01 0.000000e+00   0.00000
#> 9            QT_TEC_SUPERIOR_MASC 2.037296e+01 0.000000e+00   0.00000
#> 10      QT_TEC_ESPECIALIZACAO_FEM 2.073570e+01 0.000000e+00   0.00000
#> 11     QT_TEC_ESPECIALIZACAO_MASC 1.323448e+01 0.000000e+00   0.00000
#> 12            QT_TEC_MESTRADO_FEM 7.572712e+00 0.000000e+00   0.00000
#> 13           QT_TEC_MESTRADO_MASC 5.647467e+00 0.000000e+00   0.00000
#> 14           QT_TEC_DOUTORADO_FEM 1.871324e+00 0.000000e+00   0.00000
#> 15          QT_TEC_DOUTORADO_MASC 1.613971e+00 0.000000e+00   0.00000
#> 16        QT_PERIODICO_ELETRONICO 1.873010e+04 0.000000e+00   0.00000
#> 17            QT_LIVRO_ELETRONICO 1.137129e+04 0.000000e+00   0.00000
#> 18             VL_RECEITA_PROPRIA 1.453782e+08 0.000000e+00   0.00000
#> 19       VL_RECEITA_TRANSFERENCIA 7.863441e+07 0.000000e+00   0.00000
#> 20               VL_RECEITA_OUTRA 3.368883e+07 0.000000e+00   0.00000
#> 21     VL_DESPESA_PESSOAL_DOCENTE 6.632833e+07 4.186825e+06 267.15178
#> 22     VL_DESPESA_PESSOAL_TECNICO 3.309020e+07 0.000000e+00   0.00000
#> 23     VL_DESPESA_PESSOAL_ENCARGO 3.789554e+07 0.000000e+00   0.00000
#> 24             VL_DESPESA_CUSTEIO 4.250481e+07 0.000000e+00   0.00000
#> 25        VL_DESPESA_INVESTIMENTO 9.483621e+06 0.000000e+00   0.00000
#> 26            VL_DESPESA_PESQUISA 8.864040e+05 0.000000e+00   0.00000
#> 27               VL_DESPESA_OUTRA 2.771061e+07 0.000000e+00   0.00000
a <- sapply(X[,-c(1:6)], mean)
b <- rnp_media(X[,-c(1:6)])$media_a
all(a == b)
#> [1] TRUE