Tamanho da amostra para media

Calcula tamanho da amostra ideal para estimar um parametrp de media Considera caso em que a populacao original e finita ou infinita. N = Tamanho do universo finita Z = E o desvio do valor medio que aceitamos para alcancar o nivel de confianca desejado. Em funcao do nivel de confianca que buscamos, usaremos um valor determinado que e dado pela forma da distribuicao de Normal. Os valores mais frequentes sao: Nivel de confianca 90% -> Z=1,645 Nivel de confianca 95% -> Z=1,96 Nivel de confianca 99% -> Z=2,575 e = E a margem de erro maximo que eu quero admitir (p.e. 5%) s = E o desvio padrao amostral da base variavel de interesse na base original. Se nao puder obter da base total gerar de uma amostra. Se ainda nao tiver, faremo como raiz quadrada de N

Usage

rnp_nsample_media(N, s = NULL, z = NULL, e = NULL, infinito = FALSE)

Arguments

N

Valor do tamanho da base de dados original

s

E desvio padrao da variavel de interesse, ex. score medio.

z

E o desvio do valor medio que aceitamos para alcancar o nivel de confianca desejado. (ex. 25%)

e

E a margem de erro maximo que eu quero admitir (p.e. 5%)

infinito

se TRUE assume que a populacao e infinita e nao aplica fator de correcao

Value

vetor com tamanho da amostra n

Author

LOPES, J. L