Preparacao de dados para uso geral

Esta funcao trata dados como CPF, CNPJ, CEP, TELEFONE, etc. com base nas colunas informadas pelo usuario, para salvar no banco de dados.

Usage

rnp_prep(
  data,
  list_doc_cols = NULL,
  list_nam_cols = NULL,
  list_tel_cols = NULL,
  list_ema_cols = NULL,
  list_cep_cols = NULL,
  list_end_cols = NULL,
  list_dat_cols = NULL,
  list_num_cols = NULL,
  return_all = TRUE
)

Arguments

data

Objeto de dados para obter tipagem.

list_doc_cols

Documento: Lista de documentos. Sao aceitos CPF e CNPJ.

list_nam_cols

Nomes: Lista de campos de nomes. Aceita qualquer string.

list_tel_cols

Telefone: Lista de campos de telefone. Aceita numeros e strings.

list_ema_cols

E-mails: Lista de campos de emais. Aceita qualquer string.

list_cep_cols

CEPs: Lista de campos de CEP. Aceita numeros e strings.

list_end_cols

Enderecos: Lista de campos de enderecos. Aceita qualquer string.

list_dat_cols

Datas: Lista de campos de data. Aceita qualquer tipo de data ou string de data conhecida.

list_num_cols

Numeros: Lista de campos numericos ou que se parecam numeros. O R tratara como numerico.

return_all

Se TRUE, Retorna dados corrigods mais dados originais. Caso contrario, somente colunas afetadas.

Details

A funcao trata CPF e CNPJ validando os numeros e fixando quantidade de digitos. Para os telefones, a funcao separa DDI, DDD e Numero criando campos com sufixo nos nomes. Campos de texto sao limpos removendo-se espacos em branco ou acentos. Campos de data sao validados para o formado ano-med-dia padrao do SQL. Mais de 27 tipos de padrao de data sao checados. Veja rnp_limpa_date e parse_date_time para mais detalhes. Se todos os argumentos <list_xxx_cols> forem NULL, a funcao tenta buscar CPF/CNPJ.

See also

Examples

if (FALSE) {
list_doc_cols <- c("customer_cpf","merchant_cnpj") # NULL
list_tel_cols <- c("customer_phone") # NULL
list_cep_cols <- c("address_zipcode","merchant_address_zipcode") # NULL
list_nam_cols <- c("customer_name","card_name") # NULL
list_end_cols <- c("address_street","address_complement","address_neighbor","address_city") # NULL
list_ema_cols <- c("customer_email") # NULL
list_dat_cols <- c("customer_created_date","order_date_utc3","dt_partition") #NULL
da_fix <- rnp_prep_data(da,
                        list_doc_cols = list_doc_cols,
                        list_nam_cols = list_nam_cols,
                        list_tel_cols = list_tel_cols,
                        list_ema_cols = list_ema_cols,
                        list_cep_cols = list_cep_cols,
                        list_end_cols = list_end_cols,
                        list_dat_cols = list_dat_cols, return_all = FALSE)
}