Controle para leitura de arquivos grandes (Big Files)

Esta funcao e uma chave para passar argumentos para a funcao rnp_read

Usage

rnp_read_control(
  type = "fread",
  delim = NULL,
  dec = NULL,
  try_append = TRUE,
  sep = NULL,
  quote = "\"",
  escape_backslash = FALSE,
  escape_double = TRUE,
  col_names = TRUE,
  col_select = NULL,
  col_types = NULL,
  encoding = NULL,
  na = c("", "NA"),
  comment = "",
  trim_ws = FALSE,
  skip = 0,
  n_max = Inf,
  guess_max = min(1000, n_max),
  altrep = TRUE,
  progress = NULL,
  skip_empty_rows = TRUE,
  ...
)

Arguments

type

Nome do mecanisco de leitura. Podendo sem um em "fread","readr","vroom" ou "rio". Veja os detalhes abaixo para saber mais.

delim

Delimitador de colunas, se nao souber deixe NULL. Pode ser, ",", ";", "|" e outros conforme seu arquivo.

dec

Separador decimal.

try_append

Se TRUE (padrao) tenta apendar (empilhar) os dados. Se FALSE retorna lista de tabelas, uma para cada arquivo importado.

sep

Agrupador de campos para formatacao de numeros

quote

Regex para quotation. Serve para nao confundir com aspas que delimita strings.

escape_backslash

Utilizar contrabarra para saltar Caracteres especiais

escape_double

Se TRUE o valor '""' representara um unico \"

col_names

TRUE, FALSE, um vetor de nomes ou numeros que serao tratados como indices para selecionar colunas

col_select

Vetor de nomes ou indices a selecionar.

col_types

Se NULL trabalha com os padroes do pacote readr. Se data.table, utiliza os padroes da funao fread

encoding

Encoding do arquivo. Como padrao pega as configuracoes do Sistema Operacional

na

Tipo de valor para substituir NA

comment

Caractere a ser tratado como comentario. padrao e hashtag

trim_ws

Limpar espacos em branco a direita e esquerda das colunas

skip

numero de linhas a saltar na leitura dos dados

n_max

numero de linhas que quer extrair. Se Inf retorna o maximo de linhas

guess_max

Total de linhas por partes lidas (chunksize)

altrep

Controle dos tipos de dados na leitura por vroom

progress

Mostrar progresso da leitura dos dados

skip_empty_rows

Ignorar linhas vazias

...

Argumentos extras