Estatisticas descritivas por grupo

Calcula estatisticas descritivas por grupo. Ela recebe como entrada um data.frame o nome da variavel numerica e um vetor ou lista de nomes das variaveis que serao utilizadas como grupos. A funcao trabalha com apoio da funcao plyr::ddply e aceita muitos grupos.

Usage

rnp_summary_by(base, variavel, grupos, digits = 3)

Arguments

base

data.frame com as variaveis de entrada

variavel

o nome da variavel numerica entre aspas

grupos

lista ou vetor de nomes das estatisticas de agrupamento na ordem em que deseja obter os resultados.

digits

total de digitos para arredondar os decimais.

Value

As estatisticas de saida desta funcao sao: total (N), soma, numero de missing (Nmis), minimo, primeiro quartil (Q1), mediana (Q2), terceiro quartil (Q3), maximo, desvio pacrao (devpad) e coeficiente de variacao (cv) em um data.frame agrupadas conforme as classes das variaveis de agrupamento.

Author

LOPES, J. E.

Examples

rnp_summary_by(base = mtcars, variavel = "wt", grupos = "gear")
#>   gear  N   Soma Nmis   Min    Q1 Media Mediana    Q3   Max DevPad   IQR    cv
#> 1    3 15 58.389    0 2.465 3.450 3.893    3.73 3.958 5.424  0.833 0.508 0.214
#> 2    4 12 31.400    0 1.615 2.134 2.617    2.70 3.160 3.440  0.633 1.026 0.242
#> 3    5  5 13.163    0 1.513 2.140 2.633    2.77 3.170 3.570  0.819 1.030 0.311
rnp_summary_by(base = mtcars, variavel = "wt", grupos = c("gear","cyl"))
#>   gear cyl  N   Soma Nmis   Min    Q1 Media Mediana    Q3   Max DevPad   IQR
#> 1    3   4  1  2.465    0 2.465 2.465 2.465   2.465 2.465 2.465     NA 0.000
#> 2    3   6  2  6.675    0 3.215 3.276 3.337   3.337 3.399 3.460  0.173 0.122
#> 3    3   8 12 49.249    0 3.435 3.557 4.104   3.810 4.365 5.424  0.768 0.808
#> 4    4   4  8 19.025    0 1.615 1.910 2.378   2.260 2.872 3.190  0.601 0.963
#> 5    4   6  4 12.375    0 2.620 2.811 3.094   3.157 3.440 3.440  0.413 0.629
#> 6    5   4  2  3.653    0 1.513 1.670 1.826   1.826 1.983 2.140  0.443 0.313
#> 7    5   6  1  2.770    0 2.770 2.770 2.770   2.770 2.770 2.770     NA 0.000
#> 8    5   8  2  6.740    0 3.170 3.270 3.370   3.370 3.470 3.570  0.283 0.200
#>      cv
#> 1    NA
#> 2 0.052
#> 3 0.187
#> 4 0.253
#> 5 0.134
#> 6 0.243
#> 7    NA
#> 8 0.084
rnp_summary_by(base = mtcars, variavel = "wt",
               grupos = list("gear","cyl"), digits = 2)
#>   gear cyl  N  Soma Nmis  Min   Q1 Media Mediana   Q3  Max DevPad  IQR   cv
#> 1    3   4  1  2.46    0 2.46 2.46  2.46    2.46 2.46 2.46     NA 0.00   NA
#> 2    3   6  2  6.68    0 3.21 3.28  3.34    3.34 3.40 3.46   0.17 0.12 0.05
#> 3    3   8 12 49.25    0 3.44 3.56  4.10    3.81 4.36 5.42   0.77 0.81 0.19
#> 4    4   4  8 19.02    0 1.61 1.91  2.38    2.26 2.87 3.19   0.60 0.96 0.25
#> 5    4   6  4 12.38    0 2.62 2.81  3.09    3.16 3.44 3.44   0.41 0.63 0.13
#> 6    5   4  2  3.65    0 1.51 1.67  1.83    1.83 1.98 2.14   0.44 0.31 0.24
#> 7    5   6  1  2.77    0 2.77 2.77  2.77    2.77 2.77 2.77     NA 0.00   NA
#> 8    5   8  2  6.74    0 3.17 3.27  3.37    3.37 3.47 3.57   0.28 0.20 0.08